create a separate folder for animal shelter
