+++ python/stack_and_queue/stack_queue_animal_shelter/stack-queue-animal-shelter.png
