+++ python/stack_and_queue/stack_queue_animal_shelter/stack_queue_animal_shelter.py
